requirements.txt
python-pptx>=0.6.23
fastapi>=0.115.0
httpx>=0.25.0
lxml>=5.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
redis>=5.0.0
rq>=1.16.0
boto3>=1.35.0
psycopg[binary]>=3.1.0
tenacity>=8.2.0
uvicorn>=0.30.0
